% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_lars_cptime.R
\name{forecast_lars_cptime}
\alias{forecast_lars_cptime}
\title{Forecast Lars Computation Time}
\usage{
forecast_lars_cptime(X, T_hope = 20, message = TRUE)
}
\arguments{
\item{X}{A numeric matrix or vector representing the input data.}

\item{T_hope}{A numeric value representing the desired computation time threshold. Default is 20.}

\item{message}{A logical value. If \code{TRUE}, messages will be displayed. Default is \code{TRUE}.}
}
\value{
A numeric value representing the predicted computation time for the Lars model.
}
\description{
This function predicts the computation time for the lars model based on input data.
It also suggests whether to use glmnet or lars based on the predicted computation time.
}
\details{
The function takes an input dataset and processes it through a neural network model defined by specific weights and biases loaded from files. The network is designed to predict the computation time for the Lars model.
}
