% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnetconf-package.R
\docType{package}
\name{glmnetconf-package}
\alias{glmnetconf}
\alias{glmnetconf-package}
\title{glmnetconf: Automatic Configuration Tuning for 'glmnet' Function}
\description{
Provides an automated configuration tuning framework for Lasso regression models using the 'glmnet' package. It optimizes model configurations to minimize prediction error within a specified time budget.
}
\author{
\strong{Maintainer}: Shuhei Muroya \email{muroya.shuhei.697@s.kyushu-u.ac.jp} (\href{https://orcid.org/0009-0009-0609-8890}{ORCID})

}
\keyword{internal}
