% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_hyperparameters.R
\name{tune_hyperparameters}
\alias{tune_hyperparameters}
\title{Automatically select hyperparameters for glmnet}
\usage{
tune_hyperparameters(
  X,
  size = 1000,
  T_hope = 20,
  seed = 1,
  message = TRUE,
  line = TRUE
)
}
\arguments{
\item{X}{Input matrix.}

\item{size}{Integer, number of samples to generate pareto front for optimization.}

\item{T_hope}{A numeric value representing the desired computation time threshold. Default is 20.}

\item{seed}{An integer specifying the random seed for reproducibility. Default is 1.}

\item{message}{A logical value. If \code{TRUE}, messages will be displayed. Default is \code{TRUE}.}

\item{line}{logical(1). If TRUE, the Pareto-front points in the plot are　connected with a polyline (frontier line). If FALSE, only the points are shown. Default is TRUE.}
}
\value{
A list containing optimal \code{nlambda}, \code{thresh}, pareto front, and pareto data.
}
\description{
This function uses a neural network model to find optimal \code{nlambda} and \code{thresh} parameters based on input data.
}
