% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_lasso.R
\name{all_lasso}
\alias{all_lasso}
\title{Perform Lasso Regression Using lars and glmnet}
\usage{
all_lasso(X, y, nlambda = 100, thresh = 1e-07, seed = 1)
}
\arguments{
\item{X}{A matrix or data frame of predictor variables.}

\item{y}{A numeric vector of response variables.}

\item{nlambda}{The number of lambda values for the custom glmnet configuration. Default is 100.}

\item{thresh}{A numeric value specifying the convergence threshold for glmnet. Default is \code{1e-7}.}

\item{seed}{An integer seed for reproducibility. Default is 1.}
}
\value{
A list containing:
\item{lars}{A list with lars cross-validation results, the lars model, and coefficients for the optimal model.}
\item{glmnet_default}{A list with default glmnet cross-validation results, the glmnet model, and coefficients for the optimal model.}
\item{glmnet_set}{A list with custom glmnet cross-validation results, the glmnet model, and coefficients for the optimal model.}
\item{time}{A list with computation times for lars, default glmnet, and custom glmnet configurations.}
}
\description{
This function performs Lasso regression using both lars and glmnet, with default
and custom configurations. It compares computation times and returns the models,
cross-validation results, and coefficients.You can check the difference of three ways.
}
\details{
This function compares the performance of Lasso regression using lars and glmnet
with both default and custom configurations. For glmnet, the function allows customization
of the number of lambda values and the convergence threshold.

The function calculates the coefficients for the optimal model based on cross-validation results.
Additionally, it measures the computation time for each method.
}
\examples{
\dontrun{
# Example data
x <- matrix(rnorm(100 * 10), 100, 10)
y <- rnorm(100)

# Perform Lasso regression
result <- all_lasso(X = x, y = y)
print(result)
}

}
