% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_lasso.R
\name{auto_lasso}
\alias{auto_lasso}
\title{Lasso Regression using the better package lars or glmnet with considering its computation time}
\usage{
auto_lasso(
  X,
  y,
  new_x = NULL,
  size = 1000,
  T_hope = 20,
  seed = 1,
  message = TRUE,
  line = TRUE
)
}
\arguments{
\item{X}{A matrix or data frame of predictor variables.}

\item{y}{A numeric vector of response variables.}

\item{new_x}{Optional. A matrix or data frame for prediction. If provided, predictions will be made for \code{new_x}.}

\item{size}{An integer specifying the number of samples for the glmnet lambda search. Default is 1000.}

\item{T_hope}{A numeric value specifying the computation time threshold to decide between lars and glmnet. Default is 20.}

\item{seed}{An integer specifying the random seed for reproducibility. Default is 1.}

\item{message}{Logical. If \code{TRUE}, progress messages will be displayed. Default is \code{TRUE}.}

\item{line}{logical(1). If TRUE, the Pareto-front points in the plot are　connected with a polyline (frontier line). If FALSE, only the points are shown. Default is TRUE.}
}
\value{
A list containing the following components:
\item{method}{A character string indicating the method used ("lars" or "glmnet").}
\item{timing}{A list containing the predicted time for 'lars' and the threshold \code{T_hope}.}
\item{configuration}{A list containing the selected configuration (e.g., \code{nlambda}, \code{thresh}) if 'glmnet' is used. NULL if 'lars' is used.}
\item{cv}{The cross-validation result object (\code{cv.lars} or \code{cv.glmnet}).}
\item{model}{The fitted model object (\code{lars} or \code{glmnet}).}
\item{selection}{A list containing the best tuning parameter (\code{best_step} for 'lars', \code{best_lambda} for 'glmnet').}
\item{coefficients}{The regression coefficients of the best model.}
\item{prediction}{Predictions for \code{new_x} (if provided).}
\item{Pareto_front}{Object containing Pareto front information (only for 'glmnet').}
\item{Pareto_front_data}{Data frame of the Pareto front (only for 'glmnet').}
}
\description{
This function automatically performs Lasso regression using either lars or glmnet,
depending on the specified computation time threshold (\code{T_hope}). It selects the best
model based on cross-validation and can optionally make predictions on new data.
}
\details{
The function decides between lars and glmnet based on the computation time threshold (\code{T_hope}).
If \code{T_hope} is greater than the estimated time for lars, the function uses lars for Lasso regression.
Otherwise, glmnet is used. Cross-validation is performed to select the optimal model.
}
