% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.glmnetconf}
\alias{plot.glmnetconf}
\title{Plot Method for glmnetconf Objects}
\usage{
\method{plot}{glmnetconf}(x)
}
\arguments{
\item{x}{A \code{glmnetconf} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Plot Method for glmnetconf Objects
}
