% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.glmnetconf}
\alias{print.glmnetconf}
\title{Print Method for glmnetconf Objects}
\usage{
\method{print}{glmnetconf}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{A \code{glmnetconf} object.}

\item{digits}{The number of significant digits to use when printing Deviance.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print Method for glmnetconf Objects
}
